Ensure sysbox-fs nsenter processes don't leak host mount info inside containers.
Background:

When a container performs an access to /proc or /sys resources emulated by
sysbox-fs, or when it runs a system call trapped by Sysbox, Sysbox sometimes
dispatches an ephemeral "nsenter" process that enters the container's
namespaces, performs an action within them, and exits.

Problem:

An issue was discovered and reported by a team at Klarna.com, whereby the
nsenter process was exposing host mount info inside the container. Since it's
common for the nsenter process to enter the container's pid namespace, it's PID
and therefore it's mounts, were visible by other processes within the
container. Thus, host mount info was "leaked" inside the container, which is not
ideal (even though it's unlikely the container could do much with it since it's
guarded by it's chroot jail and linux namespaces, particularly the user
namespace).

The underlying problem was that the nsenter process was in some cases not
entering the container's mount namespace, because it wanted to perform
mounts of procfs and sysfs without the container noticing. However, by not
entering the container's mount namespaces, it was carrying with it the
host mounts it inherits from its parent process (sysbox-fs).

Solution:

The solution is for the nsenter process to always enter the container's mount
namespace; this way it let's go of it's parent process mounts and only sees the
container's mounts (i.e., no more host mount leakage in the container). If the
nsenter process needs to do some mounts of it's own, then after having entered
the container's mount namespaces, it will unshare its mount namespace so that it
can perform these mounts without affecting the container. Note that even though
other container processes can see the PID of the ephemeral nsenter process, they
won't be able to access any new mounts done by this process and won't have
permission to enter its mount namespace either.

Special thanks to Gabriele Diener at Klarna.com for reporting this issue.

Signed-off-by: Cesar Talledo <cesar.talledo@docker.com>