Fix for issue slowing down Sysbox due to accumulation of large numbers of mountpoints.
The problem is reproduced when the `--allow-immutable-unmounts=false` is set. In
this scenario, upon execution of an unmount() instruction, sysbox-fs is expected
to iterate through the list of mountpoints within the corresponding sysbox
container to decide if the unmount operation can proceed. This code path is not
executed in the default scenario (i.e., `--allow-immutable-unmounts=true`).

Another aspect required to reproduce the issue is a large number of mount/umount
instructions with targets that include at least one softlink along the target
paths (e.g., `/var/run -> /run`). In these cases, the mounted path registered in
the kernel always corresponds to the resolved version of its target (e.g.,
`/run/blah`).

The problem is reproduced when sysbox-fs processes an unmount() instruction for
an unresolved target (e.g., `/var/run/blah`), and sysbox-fs' immutable-mounts
implementation attempts to identify this target within the list of mountpoints
for this process. This unresolved target won't be ever found due to the comment
above, so sysbox-fs will return an EINVAL back to the user.

As a result, as no unmount() operation will ever succeed for these targets,
sysbox-fs will accumulate large numbers of mountpoints (specially, if docker's
health-check feature is configured every a few seconds), which will considerably
slow sysbox-fs operations within this sysbox container.

The fix is to resolve the mount targets during the mount() and umount()
operations. Technically speaking, this resolution is only required for the
umount() instruction, since the kernel already takes care of resolving each
target, but in this fix I'm resolving both instructions for consistency
purposes.

Signed-off-by: Rodny Molina <rodny.molina@docker.com>