Fix bug in nsenter mounts of procfs and sysfs.
Fix a bug in the nsenter agent when mounting and later unmounting procfs and
sysfs. The bug was causing the nsenter agent to fail to correctly unmount the
temporary procfs and sysfs mounts it sets up, causing container mounts to be
left in a bad state. For example, see the last three /proc mounts in this
container findmnt output:

|-/proc                   proc                             proc     rw,nosuid,nodev,noexec,relatime
| |-/proc/swaps           sysboxfs[/proc/swaps]            fuse     rw,nosuid,nodev,relatime,user_id=0,group_id=0,default_permissions,allow_other
| |-/proc/sys             sysboxfs[/proc/sys]              fuse     rw,nosuid,nodev,relatime,user_id=0,group_id=0,default_permissions,allow_other
| |-/proc/uptime          sysboxfs[/proc/uptime]           fuse     rw,nosuid,nodev,relatime,user_id=0,group_id=0,default_permissions,allow_other
| `-/proc                 proc                             proc     rw,nosuid,nodev,noexec,relatime
|   `-/proc               proc                             proc     rw,nosuid,nodev,noexec,relatime
|     `-/proc             proc                             proc     rw,nosuid,nodev,noexec,relatime

This is a problem because it leaves the real "procfs" inside the container
mount-ns (which in turn breaks some container workloads).

Signed-off-by: Cesar Talledo <cesar.talledo@docker.com>