Fix access to /proc/sys/kernel/shm* inside container.
Back in commit 3599e78, we determined that /proc/sys/kernel/shm* sysctls were
namespaced by the kernel, however accessing them from within a Sysbox container
was not possible because the kernel was only allowing "true root" to write to
the sysctls. Since this is clearly a kernel bug, to work-around it the
above-referenced commit added a procfs emulation handler that entered all
container namespaces except the user-namespace before accessing the
/proc/sys/kernel/shm* sysctls. We also added a test to verify the work-around
behaved as expected.

However, starting with kernel 6.9 (see upstream commit 50ec499b9a43), access to
/proc/sys/kernel/shm* from within a user-ns works (as it should be). However
this causes the work-around that we previously added in Sysbox to fail. Users
reported this in Sysbox issues 903 and 909.

To make Sysbox work, this commit causes sysbox-fs to check if the host supports
writes to /proc/sys/kernel/shm* sysctls from within a user-ns. The check is done
when sysbox-fs starts up. If the host kernel supports such access, then Sysbox
has nothing to do to enable containers to access those sysctls. If the host kernel does
not support such access, then Sysbox applies the work-around we added back in
commit 3599e78.

Signed-off-by: Cesar Talledo <cesar.talledo@docker.com>