Trap openat2 system call to allow access to sysbox-fs mounts under /proc and /sys.
The latest runc version uses the openat2() system call to ensure that there are
no overmounts on top of procfs and sysfs, to secure against malicious attacks
that allow other containers to fake procfs and sysfs via such mounts.

However this also causes problems when runc runs inside a sysbox container
(e.g., for docker-in-docker or K8s-in-docker), because sysbox emulates portions of
the container's /proc and /sys using overmounts (e.g., /proc/sys, /proc/uptime, etc.),
for legitimate reasons.

To work-around this, this commit adds a syscall trap on the openat2() system call.
When openat2() executes within a sysbox container and opens a file under /proc or
under /sys emulated by sysbox, then sysbox will ensure that syscall works correctly
by bypassing any overmounts check on the syscall. If the syscall does not target
a file emulated by sysbox, then sysbox lets the kernel process the syscall as usual.

Signed-off-by: Cesar Talledo <cesar.talledo@docker.com>